RSKT-Seg-main/RSDA-Seg-main/requirements.txt
open-clip-torch==2.24.0
scipy==1.10.1
ftfy==6.0.1
opencv-python==4.5.1.48
setuptools==69.5.1
pillow==10.3.0
imageio==2.34.1
timm==0.8.3.dev0
regex
einops
git+https://github.com/facebookresearch/detectron2.git
#git+git+https://githubfast.com/facebookresearch/detectron2.git